chore: move `MarshalDelimited*` functions
Move `MarshalDelimited*` functions to the files in which they are used
