+++ pkg/shares/shares.go
@@ -1,10 +1,7 @@
 package shares
 
 import (
-	"encoding/binary"
-
 	"github.com/celestiaorg/nmt/namespace"
-	coretypes "github.com/tendermint/tendermint/types"
 )
 
 // Share contains the raw share data (including namespace ID).
@@ -36,23 +33,4 @@ func (ns NamespacedShares) RawShares() [][]byte {
 		res[i] = nsh.Share
 	}
 	return res
-}
-
-// MarshalDelimitedTx prefixes a transaction with the length of the transaction
-// encoded as a varint.
-func MarshalDelimitedTx(tx coretypes.Tx) ([]byte, error) {
-	lenBuf := make([]byte, binary.MaxVarintLen64)
-	length := uint64(len(tx))
-	n := binary.PutUvarint(lenBuf, length)
-	return append(lenBuf[:n], tx...), nil
-}
-
-// MarshalDelimitedMessage marshals the raw share data (excluding the namespace)
-// of this message and prefixes it with the length of the message encoded as a
-// varint.
-func MarshalDelimitedMessage(msg coretypes.Message) ([]byte, error) {
-	lenBuf := make([]byte, binary.MaxVarintLen64)
-	length := uint64(len(msg.Data))
-	n := binary.PutUvarint(lenBuf, length)
-	return append(lenBuf[:n], msg.Data...), nil
 }
